Move functions about generating codes from a JSON to a new file generate_codes.py
